Внесены правки: переименованы файлы и константы согласно критериям оценивания, изменен компонент карточек
