+++ project/src/pages/favorites-page/favorites-page.tsx
