+++ project/src/pages/property-page/property-page.tsx
